moved KafkaSinkWrappingProducer & co to kafka-connect-adaptor
