+++ pulsar-io/kafka/pom.xml
@@ -61,10 +61,9 @@
     </dependency>
 
     <dependency>
-      <groupId>org.apache.kafka</groupId>
-      <artifactId>connect-file</artifactId>
-      <version>${kafka-client.version}</version>
-      <scope>test</scope>
+      <groupId>${project.groupId}</groupId>
+      <artifactId>pulsar-io-kafka-connect-adaptor</artifactId>
+      <version>${project.version}</version>
     </dependency>
 
   </dependencies>
